fixed shell scripts layout and output messages

+++ bridges.sh
@@ -30,10 +30,20 @@ print_elapsed_time() {
 	local seconds=$((delta_time % 60))
 	printf "%02d:%02d:%02d\n" "$hours" "$minutes" "$seconds"
 }
+read -rep "create new bridges output? (warning: this will override the previous output) (y/n): " create_bridges
+if ! [[ "$create_bridges" == "y" || "$create_bridges" == "Y" || "$create_bridges" == "n" || "$create_bridges" == "N" ]]
+then
+	echo "error: unrecognized input">&2
+	exit 1
+fi
+if [[ "$create_bridges" == "n" || "$create_bridges" == "N" ]]
+then
+	exit 0
+fi
 echo "* ${fg_yellow}creating${normal} output for each of the bridges"
-echo "|\\"
 mkdir -p tmp/bridges
 rm -rf tmp/bridges/*
+echo "|\\"
 echo "| * ${fg_yellow}working on${normal} Warren bridge problem"
 echo "| |\\"
 update_start_time
